Fixed matching excluded modulerefs
The moduleref values were matched against the patterns instead of the other
way around. This broke when --exclude-moduleref=foobar was passed but the
assembly had a moduleref called foobar.dll, thus it wasn't matching!

The exclude pattern also has to use ^ and $ else it will also match libs that
start or end with completely different filenames.

Also renamed $name to $moduleref which probably lead to the confusion.

(thanks goes to Chow Loong Jin for his help to sort this out)